refactoring plolty display
