+++ magpylib/_src/display/plotly/__init__.py
@@ -1,0 +1,1 @@
+"""display.plotly"""
